+++ src/lib/components/admin/Applications.svelte
@@ -13,8 +13,8 @@
     import MaterialSymbolsCheckRounded from "~icons/material-symbols/check-rounded";
     import MaterialSymbolsCloseRounded from "~icons/material-symbols/close-rounded";
     import MaterialSymbolsDeleteRounded from "~icons/material-symbols/delete-rounded";
-    import CocButton from "../CocButton.svelte";
-    import { toast } from "../toast";
+    import CocButton from "$lib/components/app/CocButton.svelte";
+    import { toast } from "$lib/components/app/toast";
     import UserName from "./UserName.svelte";
 
     interface Props {
