+++ src/lib/components/app/ClanInfo.svelte
