+++ src/routes/(admin)/admin/applications/+page.svelte
