+++ src/routes/(admin)/admin/users/+page.svelte
@@ -2,9 +2,9 @@
     import { invalidateAll } from "$app/navigation";
     import CocAccountsUsersFormsWrapper from "$lib/components/admin/wrappers/CocAccountsUsersFormsWrapper.svelte";
     import UserNameUsersFormsWrapper from "$lib/components/admin/wrappers/UserNameUsersFormsWrapper.svelte";
-    import { toast } from "$lib/components/toast";
-    import Button from "$lib/components/ui/Button.svelte";
-    import Grid from "$lib/components/ui/Grid.svelte";
+    import { toast } from "$lib/components/app/toast";
+    import Button from "$lib/components/app/ui/Button.svelte";
+    import Grid from "$lib/components/app/ui/Grid.svelte";
     import type { InsertCoc, InsertUser } from "$lib/server/schema";
     import type { GridOptions } from "@ag-grid-community/core";
     import { makeSvelteCellRenderer } from "ag-grid-svelte5-extended";
