+++ src/routes/(app)/auth/logout/+server.ts
