+++ src/routes/(app)/clans/+page.server.ts
