+++ src/routes/(app)/clans/rules/+page.server.ts
